+++ src/base/index.spec.ts
@@ -21,6 +21,8 @@ describe(schematicName, () => {
     name: 'test',
     repositoryUrl: 'test',
     license: 'test',
+    author: 'test',
+    email: 'test',
   };
 
   it('should create all files of a project', () => {