+++ src/base/schema.d.ts
@@ -11,15 +11,30 @@ export interface Schema {
    * name of the project
    */
   name: string;
+
+  /**
+   * description of the project
+   */
+  description?: string;
   
   /**
    * url of the repository
    */
-  repositoryUrl: string;
+  repositoryUrl?: string;
+
+  /**
+   * author of the project (name or company)
+   */
+  author?: string;
+
+  /**
+   * contact email of the author
+   */
+  email?: string;
   
   /**
    * type of the license
    */
-  license: string;
+  license?: string;
   
 }