+++ src/init/schema.d.ts
@@ -1,10 +1,33 @@
 export interface Schema {
+
   /**
-   * The name of the project.
+   * name of the project
    */
   name: string;
+
   /**
-   * Short description of the project.
+   * description of the project
    */
   description: string;
+
+  /**
+   * url of the repository
+   */
+  repositoryUrl?: string;
+
+  /**
+   * author of the project (name or company)
+   */
+  author?: string;
+
+  /**
+   * contact email of the author
+   */
+  email?: string;
+
+  /**
+   * type of the license
+   */
+  license?: string;
+
 }